fix: rename bolt package to badger
