+++ datastore/badger/event_delivery.go
@@ -1,4 +1,4 @@
-package bolt
+package badger
 
 import (
 	"context"
