+++ datastore/badger/event_delivery_test.go
@@ -1,4 +1,4 @@
-package bolt
+package badger
 
 import (
 	"context"
